src/Mindbox.Geometry.Domain/Figures/Circle.cs
using System;
using Mindbox.Geometry.Abstract;

namespace Mindbox.Geometry.Domain.Figures
{
    public class Circle : IFigure
    {
        private double _radius;
        public double Radius
        {
            get => _radius;
            set
            {
                if (value <= 0)
                    throw new ArgumentException(message: "Circle radius should be greater than 0");

                _radius = value;
            }
        }

        public Circle(double radius)
        {
            Radius = radius;
        }

        public double CalculateArea() => Math.PI * Math.Pow(Radius, 2);
    }
}
